requirements/base.txt
django==1.9.3
django-import-export==0.4.5